requirements.txt

PyOpenGL
numpy
Pillow
Pygame